+++ src/features/goal/components/goal/index.const.ts
@@ -1,0 +1,10 @@
+export const generateDateString = (dateString) => {
+  const date = new Date(dateString);
+
+  const formattedDate = date.toLocaleString("ko-KR", {
+    month: "long",
+    day: "numeric"
+  });
+
+  return formattedDate;
+};
